+++ packages/vchart/src/event/events/dimension/util/common.ts
@@ -2,11 +2,13 @@
 import type { IChart } from '../../../../chart/interface';
 import type { IDimensionData, IDimensionInfo } from '../interface';
 import { isNil, array, isValid, isValidNumber } from '@visactor/vutils';
+// eslint-disable-next-line no-duplicate-imports
+import type { Maybe } from '@visactor/vutils';
 import type { AxisComponent } from '../../../../component/axis/base-axis';
-import type { CoordinateType, ILayoutPoint } from '../../../../typings';
+import type { CoordinateType, Datum, ILayoutPoint } from '../../../../typings';
 import { isDiscrete } from '@visactor/vscale';
 import type { ICartesianLinearAxisSpec } from '../../../../component';
-import type { Maybe } from '@visactor/vutils';
+import type { ISeries } from '../../../../series';
 
 const isInBound = (pos: ILayoutPoint, min: ILayoutPoint, max: ILayoutPoint): boolean =>
   pos.x >= min.x && pos.x <= max.x && pos.y >= min.y && pos.y <= max.y;
@@ -60,42 +62,56 @@ export const getDimensionData = (
         const viewData = series.getViewData()?.latestData;
         if (dimensionField && viewData) {
           if (isDiscreteAxis) {
+            const datums: Datum[] = [];
+            const datumIdList: number[] = [];
+            viewData.forEach((datum: any, i: number) => {
+              if (datum[dimensionField[0]]?.toString() === value?.toString()) {
+                datums.push(datum); // 获取该维度项所对应的数据
+                datumIdList.push(i);
+              }
+            });
             data.push({
               series,
-              datum: viewData.filter(
-                (datum: any) => datum[dimensionField[0]]?.toString() === value?.toString() // 获取该维度项所对应的数据
-              )
+              datum: datums,
+              key: getDimensionDataKey(series, datumIdList)
             });
           } else {
             // 连续轴
             if (isValid(dimensionField[1])) {
               // 直方图情况，根据范围取 datum
-              data.push({
-                series,
-                datum: viewData.filter((datum: any) => {
-                  if (datum[dimensionField[0]]?.toString() === value?.toString()) {
-                    return true;
-                  }
-                  return (
-                    isValid(datum[dimensionField[0]]) &&
+              const datums: Datum[] = [];
+              const datumIdList: number[] = [];
+              viewData.forEach((datum: any, i: number) => {
+                if (
+                  datum[dimensionField[0]]?.toString() === value?.toString() ||
+                  (isValid(datum[dimensionField[0]]) &&
                     isValid(datum[dimensionField[1]]) &&
                     value >= datum[dimensionField[0]] &&
-                    value < datum[dimensionField[1]]
-                  );
-                })
+                    value < datum[dimensionField[1]])
+                ) {
+                  datums.push(datum); // 获取该维度项所对应的数据
+                  datumIdList.push(i);
+                }
+              });
+              data.push({
+                series,
+                datum: datums,
+                key: getDimensionDataKey(series, datumIdList)
               });
             } else {
               // 散点图情况，依据轴上的配置判断
               const range = (axis.getSpec() as ICartesianLinearAxisSpec).tooltipFilterRange;
               const rangeArr = (isValidNumber(range) ? [-range, range] : range) as Maybe<[number, number]>;
-              let datums: any[] = [];
+              let datums: Datum[] = [];
+              let datumIdList: number[] = [];
               if (rangeArr) {
                 // 根据范围取 datum
-                viewData.forEach((datum: any) => {
+                viewData.forEach((datum: any, i: number) => {
                   if (isValid(datum[dimensionField[0]])) {
                     const delta = datum[dimensionField[0]] - value;
                     if (delta >= rangeArr[0] && delta <= rangeArr[1]) {
                       datums.push(datum);
+                      datumIdList.push(i);
                     }
                   }
                 });
@@ -103,23 +119,26 @@ export const getDimensionData = (
                 // 根据最近距离取 datum
                 let minDelta = Infinity;
                 let deltaSign = 0;
-                viewData.forEach((datum: any) => {
+                viewData.forEach((datum: any, i: number) => {
                   if (isValid(datum[dimensionField[0]])) {
                     const delta = Math.abs(datum[dimensionField[0]] - value);
                     const sign = Math.sign(datum[dimensionField[0]] - value);
                     if (delta < minDelta) {
                       minDelta = delta;
                       datums = [datum];
+                      datumIdList = [i];
                       deltaSign = sign;
                     } else if (delta === minDelta && sign === deltaSign) {
                       datums.push(datum);
+                      datumIdList.push(i);
                     }
                   }
                 });
               }
               data.push({
                 series,
-                datum: datums
+                datum: datums,
+                key: getDimensionDataKey(series, datumIdList)
               });
             }
           }
@@ -129,6 +148,8 @@ export const getDimensionData = (
   }
   return data;
 };
+
+const getDimensionDataKey = (series: ISeries, datumIdList: number[]) => `${series.id}_${datumIdList.join('_')}`;
 
 export const getAxis = <T extends AxisComponent>(
   chart: IChart,