+++ packages/vmind/src/calculator/mocks/table.ts
@@ -1,0 +1,12 @@
+import type {
+  TableData,
+} from '../types'
+
+export const mockSimpleTable: TableData = [
+  { name: 'Alice', subject: 'Sword', score: 100, age: 11 },
+  { name: 'Eugeo', subject: 'Sword', score: 100, age: 11 },
+  { name: 'Kirito', subject: 'Sword', score: 200, age: 17 },
+  { name: 'Fanatio', subject: 'Sword', score: 100, age: 20 },
+  { name: 'Cardinal', subject: 'System', score: 500, age: 200 },
+  { name: 'Admin', subject: 'System', score: 1000, age: 300 },
+]