+++ packages/vmind/src/gpt/utils.ts
@@ -46,7 +46,6 @@ export const parseGPTJson = (JsonStr: string, prefix?: string) => {
     try {
       return JSON5.parse(str);
     } catch (err) {
-      console.info(err);
       return {
         error: true
       };
@@ -55,7 +54,8 @@ export const parseGPTJson = (JsonStr: string, prefix?: string) => {
   //解析GPT返回的JSON格式
   if (prefix) {
     //被某些字符包裹
-    const splittedStr = JsonStr.split(prefix)[1];
+    const splitArr = JsonStr.split(prefix);
+    const splittedStr = splitArr[splitArr.length - 2];
     const res = parseNoPrefixStr(splittedStr);
     if (!res.error) {
       return res;
@@ -77,4 +77,8 @@ export const parseGPTResponse = (GPTRes: LLMResponse) => {
   const content = choices[0].message.content;
   const resJson: GPTDataProcessResult = parseGPTJson(content, '```');
   return resJson;
+};
+
+export const replaceAll = (originStr: string, replaceStr: string, newStr: string) => {
+  return originStr.split(replaceStr).join(newStr);
 };